Add files via upload
This is a project I created for my foundational programming course. It serves as an introductory exercise for learning Python and combines my interest in sports. The collector is capable of gathering various data of NBA players such as their positions, heights, draft positions, salaries, etc., from hupu.com based on instructions and outputs them into an Excel file.